# docs/api-specification-oas3.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "UserRating API",
    "description": "This is the main API for the user rating data collection.",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/redis/set": {
      "post": {
        "tags": [
          "Redis controller"
        ],
        "operationId": "set",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedisForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/redis/get": {
      "get": {
        "tags": [
          "Redis controller"
        ],
        "operationId": "get",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/redis/delete": {
      "delete": {
        "tags": [
          "Redis controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RedisForm": {
        "required": [
          "key",
          "value"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "description": "API key for the organisation. <br /><br />**Please note:** that the API key needs to be inserted manually in the requests when utilizing the Swagger UI.<br /><br />",
        "name": "X-API-KEY",
        "in": "header"
      }
    }
  }
}